refactor(protocol): split typing-analytics-service into cohesive modules
